working test process

+++ PSPDataFetchLayer/Models/PspTimeValTuple.cs
@@ -8,6 +8,6 @@ namespace PSPDataFetchLayer.Models
     public class PspTimeValTuple
     {
         public int TimeInt { get; set; }
-        public double? Val { get; set; }
+        public decimal? Val { get; set; }
     }
 }
